Split reducers into separate files for clean organization
